feat(animation): add FBX animation parsing and utility functions
- Implemented animation utilities for handling FBX animations, including functions to get times for all axes, retrieve keyframe track values, interpolate rotations, and get default Euler order.
- Created curve parsers to parse animation curve nodes and animation curves from FBX objects.
- Developed layer parser to handle animation layers and assign curves to animation nodes.
- Added stack parser to connect animation stacks to their respective layers.
- Introduced track generators to create Three.js animation clips from parsed animation nodes, including support for position, rotation, scale, and morph targets.
- Updated text parser types to ensure strict typing for userData and attrValues.
