+++ src/parse/animation/index.ts
@@ -1,0 +1,18 @@
+/**
+ * FBX 动画解析模块导出
+ */
+
+// 导出曲线解析器
+export * from './curve-parsers';
+
+// 导出层解析器
+export * from './layer-parser';
+
+// 导出堆栈解析器
+export * from './stack-parser';
+
+// 导出轨道生成器
+export * from './track-generators';
+
+// 导出动画工具函数
+export * from './animation-utils';
